Welcome cats-effect-sync
