+++ sync/shared/src/main/scala/cats/effect/sync/SyncIO.scala
@@ -14,9 +14,10 @@
  * limitations under the License.
  */
 
-package cats.effect
+package cats.effect.sync
 
 import cats.{Align, Eval, Functor, Now, Show, StackSafeMonad}
+import cats.effect.kernel.{MonadCancel, Sync}
 import cats.effect.util.{ArrayStack, ByteStack}
 import cats.kernel.{Monoid, Semigroup}
 import cats.data.Ior
